Reworked ToroidalBitMatrix to be generic over a storage type
Because ToroidalBitMatrix is now parameterized over unsigned ints, it
should be possible to use packed u8s instead of u32s, meaning there will
be no need for packing/unpacking to other types.

Additionally, upon testing the old implementation was found to be
bugged. Integrating this new implementation will be required to validate
it.